// template/react-starter-kit/src/.generated/api/model/findIdItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export interface FindIdItem {
  /** 마스킹된 이메일 */
  maskedEmail: string;
  /** 가입 로그인 수단 (credential, google, kakao, naver, github 등) */
  provider: string;
  /** 가입 일시 */
  createdAt: string;
}